requirements-dashboard.txt
# Dashboard generation dependencies
psycopg2-binary>=2.9.9
duckdb>=1.0.0
pandas>=2.1.4
pyarrow>=15.0.0
plotly>=5.18.0
scikit-learn>=1.3.2
numpy>=1.26.2
mistune>=3.0.0
pyyaml>=6.0.0
